Fixed patches so they match the the current repo
As detailed in github issue #12, build currently fails. This is because the (base64 encoded) patches in the Dockerfile don't match the current repo's data. I applied the changes that the patches are supposed to do to the current repo's files by hand and created new patches that lead to successful compilation.

I suspect that this will also solve issue #11, since the second patch is supposed to do something identical to the solution proposed by maryribko (just by changing the .prototxt, not the python code). A failed patch could have caused the symptoms of getting a gradient with all 0.